requirements.txt
asgiref==3.5.1
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==4.0.4
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-storages==1.12.3
Pillow==9.1.0
python-dateutil==2.8.2
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
